+++ src/util/createServerRootMixin.js
@@ -1,7 +1,6 @@
 import instantsearch from 'instantsearch.js/es';
 import algoliaHelper from 'algoliasearch-helper';
 import { isVue3, isVue2, Vue2, createSSRApp } from '../util/vue-compat';
-import { renderToString } from '../util/renderToString';
 import { _objectSpread } from '../util/polyfills';
 const { SearchResults, SearchParameters } = algoliaHelper;
 import { warn } from './warn';
@@ -14,6 +13,35 @@ function walkIndex(indexWidget, visit) {
     visit(widget);
     walkIndex(widget, visit);
   });
+}
+
+export function renderToString(app) {
+  let _renderToString;
+  try {
+    _renderToString = isVue3
+      ? require('@vue/server-renderer').renderToString
+      : require('vue-server-renderer/basic');
+  } catch (e) {
+    // error is handled by regular if, in case it's `undefined`
+  }
+  if (!_renderToString) {
+    if (isVue3) {
+      throw new Error('you need to install @vue/server-renderer');
+    } else {
+      throw new Error('you need to install vue-server-renderer');
+    }
+  }
+
+  if (isVue3) {
+    return _renderToString(app);
+  } else {
+    return new Promise((resolve, reject) =>
+      _renderToString(app, (err, res) => {
+        if (err) reject(err);
+        resolve(res);
+      })
+    );
+  }
 }
 
 function searchOnlyWithDerivedHelpers(helper) {
