--- src/util/renderToString/index-3.js
@@ -1,1 +1,0 @@
-export { renderToString } from '@vue/server-renderer';
